1. Изменил размерность у большинства свойств с int на double. 2. Добавил вычисление безразмерных величин. 3. Выделил в отдельные интерфейсы группы свойств IFan. 4. Скорректировал ToPrint. 5. Переместил методы в один класс. 6. Изменил размерность свойства Size. Например, с "040" до 0,4 [м].

+++ Libraries/Description_of_objects/Parameters/DataCurve.cs
@@ -7,8 +7,8 @@ public record DataCurve
     public int DcIndex { get; init; }
     public required double DcVolumeFlow { get; init; }
     public required double DcTotalPressure { get; init; }
-    public required int DcSize { get; init; }
-    public required int DcImpellerRotationSpeed { get; init; }
+    public required double DcSize { get; init; }
+    public required double DcImpellerRotationSpeed { get; init; }
     public required IHumidAir DcAir { get; init; }
     public required double DcPower { get; init; }
